requirements.txt
# Add your project dependencies here
streamlit
langchain
langchain-groq
langchain-community
pypdf
sentence-transformers
chromadb
python-dotenv